fix: 수영장 상세 페이지
